+++ sql/NexoraDB/Tables/dbo.Users.sql
@@ -23,6 +23,7 @@ CREATE TABLE [dbo].[Users](
 	[twoFASecret] [nvarchar](100) NULL,
 	[locale] [nvarchar](3) NULL,
 	[LastLoginAt] [datetime] NULL,
+	[ui_prefs] [nvarchar](500) NULL,
  CONSTRAINT [PK__Users__CB9A1CDF7F8A761C] PRIMARY KEY CLUSTERED 
 (
 	[userID] ASC